test(orderService): updated the order service test
updated all the orderService test, to 100% coverage - added some addtional minor adjustments

partly resolves #5
